utils/plagiarism.py
import requests
from typing import Dict, Optional
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# You can integrate with a plagiarism checking API like Quillbot, Unicheck, etc.
# This is a placeholder for the actual implementation
PLAGIARISM_API_KEY = os.getenv("PLAGIARISM_API_KEY")
PLAGIARISM_API_URL = "https://api.plagiarismchecker.com/v1/check"  # Example URL

async def check_plagiarism(text: str) -> Dict[str, any]:
    """
    Check text for plagiarism using an external API
    
    Args:
        text: The text to check for plagiarism
        
    Returns:
        Dict containing plagiarism check results
    """
    if not PLAGIARISM_API_KEY:
        return {
            "status": "error",
            "message": "Plagiarism API key not configured"
        }
    
    try:
        headers = {
            "Authorization": f"Bearer {PLAGIARISM_API_KEY}",
            "Content-Type": "application/json"
        }
        
        payload = {
            "text": text,
            "language": "en",
            "scan_type": "web"  # or "web_and_publications"
        }
        
        response = requests.post(
            PLAGIARISM_API_URL,
            headers=headers,
            json=payload,
            timeout=30
        )
        
        if response.status_code == 200:
            return response.json()
        else:
            return {
                "status": "error",
                "message": f"Plagiarism API error: {response.status_code} - {response.text}"
            }
            
    except Exception as e:
        return {
            "status": "error",
            "message": f"Error checking plagiarism: {str(e)}"
        }
