requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-docx>=0.8.11
reportlab>=4.0.0
google-generativeai>=0.3.0
pydantic>=1.10.7
python-dotenv>=1.0.0
python-multipart>=0.0.6
docx2pdf>=0.1.8
httpx>=0.24.1
